Small changes in ft_atoi and ft_bzero

+++ ft_bzero.c
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i <= n)
@@ -24,3 +24,4 @@ void	ft_bzero(void *s, size_t n)
 		i++;
 	}
 }
+
